requirements.txt
pytelegrambotapi==4.14.0
python-dotenv==1.0.0